@@ -22,7 +22,9 @@ import 'dart:math' as math;
 import 'dart:typed_data';
 
 import 'package:crypto/crypto.dart';
+import 'package:image/image.dart' as im;
 import 'package:meta/meta.dart';
+import 'package:path_parsing/path_parsing.dart';
 import 'package:utf/utf.dart';
 import 'package:vector_math/vector_math_64.dart';
 
@@ -38,7 +40,7 @@ part 'src/border.dart';
 part 'src/catalog.dart';
 part 'src/color.dart';
 part 'src/colors.dart';
-part 'src/compatibility.dart';
+part 'src/data_types.dart';
 part 'src/document.dart';
 part 'src/encryption.dart';
 part 'src/exif.dart';
@@ -46,6 +48,8 @@ part 'src/font.dart';
 part 'src/font_descriptor.dart';
 part 'src/font_metrics.dart';
 part 'src/formxobject.dart';
+part 'src/function.dart';
+part 'src/graphic_state.dart';
 part 'src/graphics.dart';
 part 'src/image.dart';
 part 'src/info.dart';
@@ -58,8 +62,8 @@ part 'src/page.dart';
 part 'src/page_format.dart';
 part 'src/page_list.dart';
 part 'src/point.dart';
-part 'src/polygon.dart';
 part 'src/rect.dart';
+part 'src/shading.dart';
 part 'src/signature.dart';
 part 'src/stream.dart';
 part 'src/ttf_parser.dart';
