@@ -19,17 +19,17 @@ part of pdf;
 class PdfArrayObject extends PdfObject {
   PdfArrayObject(
     PdfDocument pdfDocument,
-    this.values,
-  )   : assert(values != null),
+    this.array,
+  )   : assert(array != null),
         super(pdfDocument);
 
-  final List<String> values;
+  final PdfArray array;
 
   @override
   void _writeContent(PdfStream os) {
     super._writeContent(os);
 
-    os.putStringArray(values);
+    array.output(os);
     os.putBytes(<int>[0x0a]);
   }
 }
