@@ -101,6 +101,14 @@ abstract class PdfFont extends PdfObject {
         pdfDocument, 'ZapfDingbats', 0.820, -0.143, _zapfDingbatsWidths);
   }
 
+  static const String _cannotDecodeMessage =
+      '''---------------------------------------------
+Cannot decode the string to Latin1.
+This font does not support Unicode characters.
+If you want to use strings other than Latin strings, use a TrueType (TTF) font instead.
+See https://github.com/DavBfr/dart_pdf/wiki/Fonts-Management
+---------------------------------------------''';
+
   /// The df type of the font, usually /Type1
   final String subtype;
 
@@ -119,9 +127,9 @@ abstract class PdfFont extends PdfObject {
   void _prepare() {
     super._prepare();
 
-    params['/Subtype'] = PdfStream.string(subtype);
-    params['/Name'] = PdfStream.string(name);
-    params['/Encoding'] = PdfStream.string('/WinAnsiEncoding');
+    params['/Subtype'] = PdfName(subtype);
+    params['/Name'] = PdfName(name);
+    params['/Encoding'] = const PdfName('/WinAnsiEncoding');
   }
 
   @Deprecated('Use `glyphMetrics` instead')
@@ -141,13 +149,12 @@ abstract class PdfFont extends PdfObject {
       final Uint8List chars = latin1.encode(s);
       final Iterable<PdfFontMetrics> metrics = chars.map(glyphMetrics);
       return PdfFontMetrics.append(metrics);
-    } catch (e) {
-      assert(false, '''\n---------------------------------------------
-Can not decode the string to Latin1.
-This font does not support Unicode characters.
-If you want to use strings other than Latin strings, use a TrueType (TTF) font instead.
-See https://github.com/DavBfr/dart_pdf/issues/76
----------------------------------------------''');
+    } catch (_) {
+      assert(() {
+        print(_cannotDecodeMessage);
+        return true;
+      }());
+
       rethrow;
     }
   }
@@ -163,18 +170,15 @@ See https://github.com/DavBfr/dart_pdf/issues/76
   @override
   String toString() => 'Font($fontName)';
 
-  PdfStream putText(String text) {
+  void putText(PdfStream stream, String text) {
     try {
-      return PdfStream()
-        ..putBytes(latin1.encode('('))
-        ..putTextBytes(latin1.encode(text))
-        ..putBytes(latin1.encode(')'));
-    } catch (e) {
-      assert(false, '''\n---------------------------------------------
-Can not decode the string to Latin1.
-This font does not support Unicode characters.
-If you want to use strings other than Latin strings, use a TrueType (TTF) font instead.
----------------------------------------------''');
+      PdfString(latin1.encode(text), PdfStringFormat.litteral).output(stream);
+    } catch (_) {
+      assert(() {
+        print(_cannotDecodeMessage);
+        return true;
+      }());
+
       rethrow;
     }
   }
