@@ -32,20 +32,19 @@ class PdfFontDescriptor extends PdfObject {
   void _prepare() {
     super._prepare();
 
-    params['/FontName'] = PdfStream.string('/' + ttfFont.fontName);
+    params['/FontName'] = PdfName('/' + ttfFont.fontName);
     params['/FontFile2'] = file.ref();
-    params['/Flags'] = PdfStream.intNum(ttfFont.font.unicode ? 4 : 32);
-    params['/FontBBox'] = PdfStream()
-      ..putIntArray(<int>[
-        (ttfFont.font.xMin / ttfFont.font.unitsPerEm * 1000).toInt(),
-        (ttfFont.font.yMin / ttfFont.font.unitsPerEm * 1000).toInt(),
-        (ttfFont.font.xMax / ttfFont.font.unitsPerEm * 1000).toInt(),
-        (ttfFont.font.yMax / ttfFont.font.unitsPerEm * 1000).toInt()
-      ]);
-    params['/Ascent'] = PdfStream.intNum((ttfFont.ascent * 1000).toInt());
-    params['/Descent'] = PdfStream.intNum((ttfFont.descent * 1000).toInt());
-    params['/ItalicAngle'] = PdfStream.intNum(0);
-    params['/CapHeight'] = PdfStream.intNum(10);
-    params['/StemV'] = PdfStream.intNum(79);
+    params['/Flags'] = PdfNum(ttfFont.font.unicode ? 4 : 32);
+    params['/FontBBox'] = PdfArray.fromNum(<int>[
+      (ttfFont.font.xMin / ttfFont.font.unitsPerEm * 1000).toInt(),
+      (ttfFont.font.yMin / ttfFont.font.unitsPerEm * 1000).toInt(),
+      (ttfFont.font.xMax / ttfFont.font.unitsPerEm * 1000).toInt(),
+      (ttfFont.font.yMax / ttfFont.font.unitsPerEm * 1000).toInt()
+    ]);
+    params['/Ascent'] = PdfNum((ttfFont.ascent * 1000).toInt());
+    params['/Descent'] = PdfNum((ttfFont.descent * 1000).toInt());
+    params['/ItalicAngle'] = const PdfNum(0);
+    params['/CapHeight'] = const PdfNum(10);
+    params['/StemV'] = const PdfNum(79);
   }
 }
