@@ -35,7 +35,7 @@ class PdfObjectStream extends PdfObject {
   /// defines if the stream needs to be converted to ascii85
   final bool isBinary;
 
-  List<int> _data;
+  Uint8List _data;
 
   @override
   void _prepare() {
@@ -45,21 +45,29 @@ class PdfObjectStream extends PdfObject {
       // The data is already in the right format
       _data = buf.output();
     } else if (pdfDocument.deflate != null) {
-      _data = pdfDocument.deflate(buf.output());
-      params['/Filter'] = PdfStream.string('/FlateDecode');
-    } else if (isBinary) {
-      // This is a Ascii85 stream
-      final Ascii85Encoder e = Ascii85Encoder();
-      _data = e.convert(buf.output());
-      params['/Filter'] = PdfStream.string('/ASCII85Decode');
-    } else {
-      // This is a non-deflated stream
-      _data = buf.output();
+      final Uint8List original = buf.output();
+      final Uint8List newData = pdfDocument.deflate(original);
+      if (newData.lengthInBytes < original.lengthInBytes) {
+        params['/Filter'] = const PdfName('/FlateDecode');
+        _data = newData;
+      }
+    }
+
+    if (_data == null) {
+      if (isBinary) {
+        // This is a Ascii85 stream
+        final Ascii85Encoder e = Ascii85Encoder();
+        _data = e.convert(buf.output());
+        params['/Filter'] = const PdfName('/ASCII85Decode');
+      } else {
+        // This is a non-deflated stream
+        _data = buf.output();
+      }
     }
     if (pdfDocument.encryption != null) {
       _data = pdfDocument.encryption.encrypt(_data, this);
     }
-    params['/Length'] = PdfStream.intNum(_data.length);
+    params['/Length'] = PdfNum(_data.length);
   }
 
   @override
