@@ -41,11 +41,26 @@ class PdfPage extends PdfObject {
   /// -1 indicates no thumbnail.
   PdfObject thumbnail;
 
+  /// Isolated transparency: If this flag is true, objects within the group
+  /// shall be composited against a fully transparent initial backdrop;
+  /// if false, they shall be composited against the group’s backdrop
+  bool isolatedTransparency = false;
+
+  /// Whether the transparency group is a knockout group.
+  /// If this flag is false, later objects within the group shall be composited
+  /// with earlier ones with which they overlap; if true, they shall be
+  /// composited with the group’s initial backdrop and shall overwrite any
+  /// earlier overlapping objects.
+  bool knockoutTransparency = false;
+
   /// This holds any Annotations contained within this page.
   List<PdfAnnot> annotations = <PdfAnnot>[];
 
   /// The fonts associated with this page
   final Map<String, PdfFont> fonts = <String, PdfFont>{};
+
+  /// The fonts associated with this page
+  final Map<String, PdfShading> shading = <String, PdfShading>{};
 
   /// The xobjects or other images in the pdf
   final Map<String, PdfXObject> xObjects = <String, PdfXObject>{};
@@ -82,8 +97,8 @@ class PdfPage extends PdfObject {
     params['/Parent'] = pdfDocument.pdfPageList.ref();
 
     // the /MediaBox for the page size
-    params['/MediaBox'] = PdfStream()
-      ..putNumArray(<double>[0, 0, pageFormat.width, pageFormat.height]);
+    params['/MediaBox'] =
+        PdfArray.fromNum(<double>[0, 0, pageFormat.width, pageFormat.height]);
 
     // Rotation (if not zero)
 //        if(rotate!=0) {
@@ -95,27 +110,52 @@ class PdfPage extends PdfObject {
     // the /Contents pages object
     if (contents.isNotEmpty) {
       if (contents.length == 1) {
-        params['/Contents'] = contents[0].ref();
+        params['/Contents'] = contents.first.ref();
       } else {
-        params['/Contents'] = PdfStream()..putObjectArray(contents);
+        params['/Contents'] = PdfArray.fromObjects(contents);
       }
     }
 
     // Now the resources
     /// This holds any resources for this page
-    final Map<String, PdfStream> resources = <String, PdfStream>{};
+    final PdfDict resources = PdfDict();
+
+    resources['/ProcSet'] = PdfArray(const <PdfName>[
+      PdfName('/PDF'),
+      PdfName('/Text'),
+      PdfName('/ImageB'),
+      PdfName('/ImageC'),
+    ]);
 
     // fonts
     if (fonts.isNotEmpty) {
-      resources['/Font'] = PdfStream()..putObjectDictionary(fonts);
+      resources['/Font'] = PdfDict.fromObjectMap(fonts);
+    }
+
+    // shading
+    if (shading.isNotEmpty) {
+      resources['/Shading'] = PdfDict.fromObjectMap(shading);
     }
 
     // Now the XObjects
     if (xObjects.isNotEmpty) {
-      resources['/XObject'] = PdfStream()..putObjectDictionary(xObjects);
+      resources['/XObject'] = PdfDict.fromObjectMap(xObjects);
     }
 
-    params['/Resources'] = PdfStream.dictionary(resources);
+    if (pdfDocument.hasGraphicStates) {
+      // Declare Transparency Group settings
+      params['/Group'] = PdfDict(<String, PdfDataType>{
+        '/Type': const PdfName('/Group'),
+        '/S': const PdfName('/Transparency'),
+        '/CS': const PdfName('/DeviceRGB'),
+        '/I': PdfBool(isolatedTransparency),
+        '/K': PdfBool(knockoutTransparency),
+      });
+
+      resources['/ExtGState'] = pdfDocument.graphicStates.ref();
+    }
+
+    params['/Resources'] = resources;
 
     // The thumbnail
     if (thumbnail != null) {
@@ -124,7 +164,7 @@ class PdfPage extends PdfObject {
 
     // The /Annots object
     if (annotations.isNotEmpty) {
-      params['/Annots'] = PdfStream()..putObjectArray(annotations);
+      params['/Annots'] = PdfArray.fromObjects(annotations);
     }
   }
 }
