@@ -19,198 +19,55 @@
 part of pdf;
 
 class PdfStream {
-  static const int precision = 5;
-  final List<int> _stream = <int>[];
+  static const int _grow = 65536;
 
-  void putStream(PdfStream s) {
-    _stream.addAll(s._stream);
+  Uint8List _stream = Uint8List(_grow);
+
+  int _offset = 0;
+
+  void _ensureCapacity(int size) {
+    if (_stream.length - _offset >= size) {
+      return;
+    }
+
+    final int newSize = _offset + size + _grow;
+    final Uint8List newBuffer = Uint8List(newSize);
+    newBuffer.setAll(0, _stream);
+    _stream = newBuffer;
   }
 
-  void putString(String s) {
-    for (int codeUnit in s.codeUnits) {
-      if (codeUnit <= 0x7f) {
-        _stream.add(codeUnit);
-      } else {
-        _stream.add(0x20);
-      }
-    }
-  }
-
-  static PdfStream string(String s) => PdfStream()..putString(s);
-
-  void putStringUtf16(String s) {
-    for (int codeUnit in s.codeUnits) {
-      _stream.add(codeUnit & 0xff);
-      _stream.add((codeUnit >> 8) & 0xff);
-    }
+  void putByte(int s) {
+    _ensureCapacity(1);
+    _stream[_offset++] = s;
   }
 
   void putBytes(List<int> s) {
-    _stream.addAll(s);
+    _ensureCapacity(s.length);
+    _stream.setAll(_offset, s);
+    _offset += s.length;
   }
 
-  void putNum(double d) {
-    assert(d != double.infinity);
-    putString(d.toStringAsFixed(precision));
+  void setBytes(int offset, Iterable<int> iterable) {
+    _stream.setAll(offset, iterable);
   }
 
-  void putNumList(List<double> d) {
-    putString(d.map((double v) {
-      assert(v != double.infinity);
-      return v.toStringAsFixed(precision);
-    }).join(' '));
+  void putStream(PdfStream s) {
+    putBytes(s._stream);
   }
 
-  void putIntList(List<int> d) {
-    putString(d.map((int v) => v.toString()).join(' '));
-  }
+  int get offset => _offset;
 
-  static PdfStream num(double d) => PdfStream()..putNum(d);
-  static PdfStream intNum(int i) => PdfStream()..putString(i.toString());
+  Uint8List output() => _stream.sublist(0, _offset);
 
-  /// Escape special characters
-  /// \ddd Character code ddd (octal)
-  void putTextBytes(List<int> s) {
-    for (int c in s) {
-      switch (c) {
-        case 0x0a: // \n Line feed (LF)
-          _stream.add(0x5c);
-          _stream.add(0x6e);
-          break;
-        case 0x0d: // \r Carriage return (CR)
-          _stream.add(0x5c);
-          _stream.add(0x72);
-          break;
-        case 0x09: // \t Horizontal tab (HT)
-          _stream.add(0x5c);
-          _stream.add(0x74);
-          break;
-        case 0x08: // \b Backspace (BS)
-          _stream.add(0x5c);
-          _stream.add(0x62);
-          break;
-        case 0x0c: // \f Form feed (FF)
-          _stream.add(0x5c);
-          _stream.add(0x66);
-          break;
-        case 0x28: // \( Left parenthesis
-          _stream.add(0x5c);
-          _stream.add(0x28);
-          break;
-        case 0x29: // \) Right parenthesis
-          _stream.add(0x5c);
-          _stream.add(0x29);
-          break;
-        case 0x5c: // \\ Backslash
-          _stream.add(0x5c);
-          _stream.add(0x5c);
-          break;
-        default:
-          _stream.add(c);
+  void putString(String s) {
+    assert(() {
+      for (final int codeUnit in s.codeUnits) {
+        if (codeUnit > 0x7f) {
+          return false;
+        }
       }
-    }
+      return true;
+    }());
+    putBytes(s.codeUnits);
   }
-
-  void putText(String s) {
-    putBytes(latin1.encode('('));
-    putTextBytes(latin1.encode(s));
-    putBytes(latin1.encode(')'));
-  }
-
-  void putLiteral(String s) {
-    putBytes(latin1.encode('('));
-    putBytes(<int>[0xfe, 0xff]);
-    putTextBytes(encodeUtf16be(s));
-    putBytes(latin1.encode(')'));
-  }
-
-  void putBool(bool value) {
-    putString(value ? 'true' : 'false');
-  }
-
-  /// Returns the ASCII/Unicode code unit corresponding to the hexadecimal digit
-  /// [digit].
-  int _codeUnitForDigit(int digit) =>
-      digit < 10 ? digit + 0x30 : digit + 0x61 - 10;
-
-  void putBinary(List<int> s) {
-    _stream.add(0x3c);
-    for (int byte in s) {
-      _stream.add(_codeUnitForDigit((byte & 0xF0) >> 4));
-      _stream.add(_codeUnitForDigit(byte & 0x0F));
-    }
-    _stream.add(0x3e);
-  }
-
-  static PdfStream binary(List<int> s) => PdfStream()..putBinary(s);
-
-  void putArray(List<PdfStream> values) {
-    putString('[');
-    for (PdfStream val in values) {
-      putStream(val);
-      putString(' ');
-    }
-    putString(']');
-  }
-
-  void putObjectArray(List<PdfObject> values) {
-    putString('[');
-    for (PdfObject val in values) {
-      putStream(val.ref());
-      putString(' ');
-    }
-    putString(']');
-  }
-
-  void putStringArray(List<String> values) {
-    putString('[' + values.join(' ') + ']');
-  }
-
-  void putDate(DateTime date) {
-    final DateTime utcDate = date.toUtc();
-    final String year = utcDate.year.toString().padLeft(4, '0');
-    final String month = utcDate.month.toString().padLeft(2, '0');
-    final String day = utcDate.day.toString().padLeft(2, '0');
-    final String hour = utcDate.hour.toString().padLeft(2, '0');
-    final String minute = utcDate.minute.toString().padLeft(2, '0');
-    final String second = utcDate.second.toString().padLeft(2, '0');
-    putText('D:$year$month$day$hour$minute${second}Z');
-  }
-
-  void putNumArray(List<double> values) {
-    putString('[');
-    putNumList(values);
-    putString(']');
-  }
-
-  void putIntArray(List<int> values) {
-    putString('[');
-    putIntList(values);
-    putString(']');
-  }
-
-  static PdfStream array(List<PdfStream> values) =>
-      PdfStream()..putArray(values);
-
-  void putDictionary(Map<String, PdfStream> values) {
-    putString('<< ');
-    values.forEach((String k, PdfStream v) {
-      putString('$k ');
-      putStream(v);
-      putString('\n');
-    });
-    putString('>>');
-  }
-
-  static PdfStream dictionary(Map<String, PdfStream> values) =>
-      PdfStream()..putDictionary(values);
-
-  void putObjectDictionary(Map<String, PdfObject> values) {
-    putDictionary(values.map((String string, PdfObject object) =>
-        MapEntry<String, PdfStream>(string, object.ref())));
-  }
-
-  int get offset => _stream.length;
-
-  List<int> output() => _stream;
 }
