@@ -17,35 +17,14 @@
 // ignore_for_file: omit_local_variable_types
 
 import 'dart:io';
-import 'dart:math' as math;
-import 'dart:typed_data';
 
-import 'package:test/test.dart';
 import 'package:pdf/pdf.dart';
 import 'package:pdf/widgets.dart';
+import 'package:test/test.dart';
+
+import 'utils.dart';
 
 Document pdf;
-
-PdfImage generateBitmap(PdfDocument pdf, int w, int h) {
-  final Uint32List bm = Uint32List(w * h);
-  final double dw = w.toDouble();
-  final double dh = h.toDouble();
-  for (int y = 0; y < h; y++) {
-    for (int x = 0; x < w; x++) {
-      bm[y * w + x] = (math.sin(x / dw) * 256).toInt() |
-          (math.sin(y / dh) * 256).toInt() << 8 |
-          (math.sin(x / dw * y / dh) * 256).toInt() << 16 |
-          0xff000000;
-    }
-  }
-
-  return PdfImage(
-    pdf,
-    image: bm.buffer.asUint8List(),
-    width: w,
-    height: h,
-  );
-}
 
 void main() {
   setUpAll(() {
@@ -154,6 +133,94 @@ void main() {
             ),
           ],
         ),
+      ),
+    ));
+  });
+
+  test('Container Widgets LinearGradient', () {
+    pdf.addPage(Page(
+      build: (Context context) => Container(
+        alignment: Alignment.center,
+        margin: const EdgeInsets.all(30),
+        padding: const EdgeInsets.all(20),
+        decoration: const BoxDecoration(
+            borderRadius: 20,
+            gradient: LinearGradient(
+              colors: <PdfColor>[
+                PdfColors.blue,
+                PdfColors.red,
+                PdfColors.yellow,
+              ],
+              begin: Alignment.bottomLeft,
+              end: Alignment.topRight,
+              stops: <double>[0, .8, 1.0],
+              tileMode: TileMode.clamp,
+            ),
+            border: BoxBorder(
+              color: PdfColors.blue800,
+              top: true,
+              left: true,
+              right: true,
+              bottom: true,
+              width: 2,
+            )),
+        width: 200,
+        height: 400,
+      ),
+    ));
+  });
+
+  test('Container Widgets RadialGradient', () {
+    pdf.addPage(Page(
+      build: (Context context) => Container(
+        alignment: Alignment.center,
+        margin: const EdgeInsets.all(30),
+        padding: const EdgeInsets.all(20),
+        decoration: const BoxDecoration(
+            borderRadius: 20,
+            gradient: RadialGradient(
+              colors: <PdfColor>[
+                PdfColors.blue,
+                PdfColors.red,
+                PdfColors.yellow,
+              ],
+              stops: <double>[0.0, .2, 1.0],
+              center: FractionalOffset(.7, .2),
+              focal: FractionalOffset(.7, .45),
+              focalRadius: 1,
+            ),
+            border: BoxBorder(
+              color: PdfColors.blue800,
+              top: true,
+              left: true,
+              right: true,
+              bottom: true,
+              width: 2,
+            )),
+        width: 200,
+        height: 400,
+        // child: Placeholder(),
+      ),
+    ));
+  });
+
+  test('Container Widgets BoxShadow', () {
+    pdf.addPage(Page(
+      build: (Context context) => Container(
+        margin: const EdgeInsets.all(30),
+        padding: const EdgeInsets.all(20),
+        decoration: const BoxDecoration(
+          boxShadow: <BoxShadow>[
+            BoxShadow(
+              blurRadius: 4,
+              spreadRadius: 10,
+              offset: PdfPoint(2, 2),
+            ),
+          ],
+          color: PdfColors.blue,
+        ),
+        width: 200,
+        height: 400,
       ),
     ));
   });
