@@ -14,22 +14,11 @@
  * limitations under the License.
  */
 
-library printing;
-
-import 'dart:async';
-import 'dart:typed_data';
-import 'dart:ui' as ui;
-
-import 'package:flutter/rendering.dart';
-import 'package:flutter/services.dart';
-import 'package:flutter/widgets.dart';
-import 'package:pdf/pdf.dart';
-import 'package:pdf/widgets.dart';
-
-part 'src/asset_utils.dart';
-part 'src/print_job.dart';
-part 'src/printer.dart';
-part 'src/printing.dart';
-part 'src/printing_info.dart';
-part 'src/raster.dart';
-part 'src/widgets.dart';
+export 'src/asset_utils.dart';
+export 'src/callback.dart';
+export 'src/pdf_preview.dart';
+export 'src/printer.dart';
+export 'src/printing.dart';
+export 'src/printing_info.dart';
+export 'src/raster.dart';
+export 'src/widget_wrapper.dart';
