@@ -14,20 +14,34 @@
  * limitations under the License.
  */
 
-part of printing;
+import 'dart:async';
+import 'dart:typed_data';
 
-class _PrintJob {
-  _PrintJob({
+import 'callback.dart';
+import 'raster.dart';
+
+/// Represents a print job to communicate with the platform implementation
+class PrintJob {
+  /// Create a print job
+  PrintJob({
     this.onLayout,
     this.onHtmlRendered,
     this.onCompleted,
     this.onPageRasterized,
   });
 
+  /// Callback used when calling Printing.layoutPdf()
   final LayoutCallback onLayout;
-  final Completer<List<int>> onHtmlRendered;
+
+  /// Callback used when calling Printing.convertHtml()
+  final Completer<Uint8List> onHtmlRendered;
+
+  /// Future triggered when the job is done
   final Completer<bool> onCompleted;
+
+  /// Stream of rasterized pages
   final StreamController<PdfRaster> onPageRasterized;
 
+  /// The Job number
   int index;
 }
